+++ nodes/WebCrawler/index.ts
@@ -1,3 +1,0 @@
-import { WebCrawler } from './WebCrawler.node';
-
-export { WebCrawler };