+++ nodes/index.ts
@@ -1,3 +1,0 @@
-import { WebCrawler } from './WebCrawler/WebCrawler.node';
-
-export { WebCrawler };